# perf(core): use FuturesUnordered instead of JoinSet