--- a/core/runtime/jsrealm.rs
+++ b/core/runtime/jsrealm.rs
@@ -46,8 +46,10 @@ impl Hasher for IdentityHasher {
 
 /// We will be experimenting with different driver types in the future. This allows us to
 /// swap the driver out for experimentation.
-#[cfg(feature = "op_driver_joinset")]
-type DefaultOpDriver = super::op_driver::JoinSetDriver;
+// #[cfg(feature = "op_driver_joinset")]
+// type DefaultOpDriver = super::op_driver::JoinSetDriver;
+
+type DefaultOpDriver = super::op_driver::FuturesUnorderedDriver;
 
 #[derive(Default)]
 pub(crate) struct ContextState<OpDriverImpl: OpDriver = DefaultOpDriver> {