--- a/core/runtime/jsrealm.rs
+++ b/core/runtime/jsrealm.rs
@@ -11,6 +11,7 @@ use crate::modules::ModuleMap;
 use crate::ops::OpCtx;
 use crate::tasks::V8TaskSpawnerFactory;
 use crate::web_timeout::WebTimers;
+use crate::GetErrorClassFn;
 use anyhow::Error;
 use futures::stream::StreamExt;
 use std::cell::Cell;
@@ -51,7 +52,6 @@ impl Hasher for IdentityHasher {
 
 type DefaultOpDriver = super::op_driver::FuturesUnorderedDriver;
 
-#[derive(Default)]
 pub(crate) struct ContextState<OpDriverImpl: OpDriver = DefaultOpDriver> {
   pub(crate) task_spawner_factory: Arc<V8TaskSpawnerFactory>,
   pub(crate) timers: WebTimers<(v8::Global<v8::Function>, u32)>,
@@ -68,6 +68,31 @@ pub(crate) struct ContextState<OpDriverImpl: OpDriver = DefaultOpDriver> {
   pub(crate) isolate: Option<*mut v8::OwnedIsolate>,
   pub(crate) exception_state: Rc<ExceptionState>,
   pub(crate) has_next_tick_scheduled: Cell<bool>,
+  pub(crate) get_error_class_fn: GetErrorClassFn,
+  pub(crate) ops_with_metrics: Vec<bool>,
+}
+
+impl<O: OpDriver> ContextState<O> {
+  pub(crate) fn new(
+    isolate_ptr: *mut v8::OwnedIsolate,
+    get_error_class_fn: GetErrorClassFn,
+    ops_with_metrics: Vec<bool>,
+  ) -> Self {
+    Self {
+      isolate: Some(isolate_ptr),
+      get_error_class_fn,
+      ops_with_metrics,
+      exception_state: Default::default(),
+      has_next_tick_scheduled: Default::default(),
+      js_event_loop_tick_cb: Default::default(),
+      js_wasm_streaming_cb: Default::default(),
+      op_ctxs: Default::default(),
+      pending_ops: Default::default(),
+      task_spawner_factory: Default::default(),
+      timers: Default::default(),
+      unrefed_ops: Default::default(),
+    }
+  }
 }
 
 /// A representation of a JavaScript realm tied to a [`JsRuntime`], that allows