--- a/core/runtime/op_driver/futures_unordered_driver.rs
+++ b/core/runtime/op_driver/futures_unordered_driver.rs
@@ -2,9 +2,6 @@
 use super::future_arena::FutureAllocation;
 use super::future_arena::FutureArena;
 use super::op_results::*;
-use super::submission_queue::new_submission_queue;
-use super::submission_queue::SubmissionQueue;
-use super::submission_queue::SubmissionQueueResults;
 use super::OpDriver;
 use crate::GetErrorClassFn;
 use crate::OpId;
@@ -17,6 +14,7 @@ use futures::future::poll_fn;
 use futures::stream::FuturesUnordered;
 use futures::task::noop_waker_ref;
 use futures::FutureExt;
+use futures::Stream;
 use std::cell::Cell;
 use std::cell::RefCell;
 use std::collections::VecDeque;
@@ -24,6 +22,7 @@ use std::future::ready;
 use std::future::Future;
 use std::pin::Pin;
 use std::rc::Rc;
+use std::task::ready;
 use std::task::Context;
 use std::task::Poll;
 
@@ -63,9 +62,8 @@ pub struct FuturesUnorderedDriver<
 
 impl<C: OpMappingContext + 'static> Drop for FuturesUnorderedDriver<C> {
   fn drop(&mut self) {
-    match self.task.take() {
-      MaybeTask::Handle(h) => h.abort(),
-      _ => {}
+    if let MaybeTask::Handle(h) = self.task.take() {
+      h.abort()
     }
   }
 }
@@ -266,7 +264,7 @@ impl<C: OpMappingContext> OpDriver<C> for FuturesUnorderedDriver<C> {
   ) -> Poll<(PromiseId, OpId, bool, OpResult<C>)> {
     let mut ops = self.completed_ops.borrow_mut();
     if ops.is_empty() {
-      self.completed_waker.register(&cx.waker());
+      self.completed_waker.register(cx.waker());
       return Poll::Pending;
     }
     let item = ops.pop_front().unwrap();
@@ -278,4 +276,80 @@ impl<C: OpMappingContext> OpDriver<C> for FuturesUnorderedDriver<C> {
   fn len(&self) -> usize {
     self.queue.len()
   }
+}
+
+impl<F: Future<Output = R>, R> SubmissionQueueFutures for FuturesUnordered<F> {
+  type Future = F;
+  type Output = F::Output;
+
+  fn len(&self) -> usize {
+    self.len()
+  }
+
+  fn spawn(&mut self, f: Self::Future) {
+    self.push(f)
+  }
+
+  fn poll_next_unpin(&mut self, cx: &mut Context) -> Poll<Self::Output> {
+    Poll::Ready(ready!(Pin::new(self).poll_next(cx)).unwrap())
+  }
+}
+
+#[derive(Default)]
+struct Queue<F: SubmissionQueueFutures> {
+  queue: RefCell<F>,
+  item_waker: UnsyncWaker,
+}
+
+pub trait SubmissionQueueFutures: Default {
+  type Future: Future<Output = Self::Output>;
+  type Output;
+
+  fn len(&self) -> usize;
+  fn spawn(&mut self, f: Self::Future);
+  fn poll_next_unpin(&mut self, cx: &mut Context) -> Poll<Self::Output>;
+}
+
+pub struct SubmissionQueueResults<F: SubmissionQueueFutures> {
+  queue: Rc<Queue<F>>,
+}
+
+impl<F: SubmissionQueueFutures> SubmissionQueueResults<F> {
+  pub fn poll_next_unpin(&mut self, cx: &mut Context) -> Poll<F::Output> {
+    let mut queue = self.queue.queue.borrow_mut();
+    self.queue.item_waker.register(cx.waker());
+    if queue.len() == 0 {
+      return Poll::Pending;
+    }
+    queue.poll_next_unpin(cx)
+  }
+}
+
+pub struct SubmissionQueue<F: SubmissionQueueFutures> {
+  queue: Rc<Queue<F>>,
+}
+
+impl<F: SubmissionQueueFutures> SubmissionQueue<F> {
+  pub fn spawn(&self, f: F::Future) {
+    self.queue.queue.borrow_mut().spawn(f);
+    self.queue.item_waker.wake_by_ref();
+  }
+
+  pub fn len(&self) -> usize {
+    self.queue.queue.borrow().len()
+  }
+}
+
+/// Create a [`SubmissionQueue`] and [`SubmissionQueueResults`] that allow for submission of tasks
+/// and reception of task results. We may add work to the [`SubmissionQueue`] from any task, and the
+/// [`SubmissionQueueResults`] will be polled from a single location.
+pub fn new_submission_queue<F: SubmissionQueueFutures>(
+) -> (SubmissionQueue<F>, SubmissionQueueResults<F>) {
+  let queue: Rc<Queue<F>> = Default::default();
+  (
+    SubmissionQueue {
+      queue: queue.clone(),
+    },
+    SubmissionQueueResults { queue },
+  )
 }