--- a/core/runtime/op_driver/futures_unordered_driver.rs
+++ b/core/runtime/op_driver/futures_unordered_driver.rs
@@ -6,7 +6,6 @@ use super::submission_queue::new_submission_queue;
 use super::submission_queue::SubmissionQueue;
 use super::submission_queue::SubmissionQueueResults;
 use super::OpDriver;
-use crate::ops::OpCtx;
 use crate::GetErrorClassFn;
 use crate::OpId;
 use crate::PromiseId;
@@ -174,14 +173,15 @@ impl<C: OpMappingContext> OpDriver<C> for FuturesUnorderedDriver<C> {
     const DEFERRED: bool,
   >(
     &self,
-    ctx: &OpCtx,
+    op_id: OpId,
+    metrics_enabled: bool,
+    get_class: GetErrorClassFn,
     promise_id: i32,
     op: impl Future<Output = Result<R, E>> + 'static,
     rv_map: C::MappingFn<R>,
   ) -> Option<Result<R, E>> {
     {
-      let info = PendingOpInfo(promise_id, ctx.id, ctx.metrics_enabled());
-      let get_class = ctx.get_error_class_fn;
+      let info = PendingOpInfo(promise_id, op_id, metrics_enabled);
 
       if LAZY {
         self.spawn_unpolled(op, move |r| {
@@ -217,13 +217,14 @@ impl<C: OpMappingContext> OpDriver<C> for FuturesUnorderedDriver<C> {
     const DEFERRED: bool,
   >(
     &self,
-    ctx: &OpCtx,
+    op_id: OpId,
+    metrics_enabled: bool,
     promise_id: i32,
     op: impl Future<Output = R> + 'static,
     rv_map: C::MappingFn<R>,
   ) -> Option<R> {
     {
-      let info = PendingOpInfo(promise_id, ctx.id, ctx.metrics_enabled());
+      let info = PendingOpInfo(promise_id, op_id, metrics_enabled);
       if LAZY {
         self.spawn_unpolled(op, move |r| {
           Self::pending_op_success(info, rv_map, r)