--- a/testing/checkin/runner/ops_async.rs
+++ b/testing/checkin/runner/ops_async.rs
@@ -2,8 +2,6 @@
 use deno_core::op2;
 
 #[op2(async)]
-pub async fn op_async_void_deferred(
-) {
+pub async fn op_async_void_deferred() {
   tokio::task::yield_now().await
 }
-